[feat] add automatic mapping of flow flags to module flags to registry loader
Since each objects contains all information for workflow and modules
flags, it makes sense to automatically create flag mappings for
conversion. Explicit definitions in the YAML take precendent over
automated mappings. There is now also a validation to make sure that a
flag_map (can be an empty dict) is present in the YAML config. Lastly,
if the ModuleType is MODULE, the flags for the module will be used
directly instead of translating from the workflow maps as the assumption
is that the module is being run as a standalone module.