[fix] fix FileNotFound error and incorrect location of reg_command.log
The CLARITY Allen registration module would fail if 'miracl reg clar_allen' was not executed in the same path that the reg_final and clar_allen_reg folders were created. This was fixed by correctly setting the absolute path the file in question. Same for the reg_command.log file.