[feat] add import API
This adds an actual import API to MIRACL. This is the first implementation. More work has to be done on public vs private methods (as far as that is possible in Python) and no unit tests have been written yet. The API is grouped into a core part and an extended part. Core methods don't have any dependencies on extended methods.