[feat] finish first version of Waxholm integration and SimpleITK Otsu masking
Also added choosing padding percentage dynamically based on used Atlas. It might be a good idea to expose the padding percentage as a flag to the user in a future version. This version of the CLARITY registration script expects the Waxholm atlas files to be available under /code/atlases/waxholm.