[fix] fix wrapped_unner closure issue
This was a nasty bug to troubleshoot. The yaml_flag_map would always be set to whatever module template was registered last. The reason was that wrapped_runner closes over a loop variablle (flag_map) that was reassigned at every iteration. Since Python closures catpure names and not values, that name resolved to the last value assigned in the loop when the runner was eventually called. In other words, wrapped_runner kept a reference to the variable name flag_map, not its value. That meant that later when the registry class method run() was calling the wrapped_runner, Python resolved flag_map to the last assigned value at function call.

This was also misleading since pdb would print different ideas for id(flag_map) which didn't contradict the actual bug. The dict object changed, but the variable name stayed the same.

The solution was to freeze the flag_map value in the wrapped_runner fn like so:

def wrapped_runner(
    script: str,
    mapping: Dict[str, object],
    *,
    _flag_map=flag_map,
    _execute=execute,
):
    return runner_func(script, mapping, flag_map=_flag_map, execute=_execute)

This guaranteed that each wrapped_runner got its own copy of the reference which prevented late binding, mutation, overwriting and the YAML order bug.

The problem here was that the code looked correct, no exceptions were thrown, pdb showed the correct values - in other words, everything broke silently.

I have to go back in the commit history and trace back where the bug was introduced and why I missed it. However, for now it's working again.